requirements.txt
colorama
discord.py
python-dotenv
pygame
aiofiles
aiohttp
aiogram
aiogram-media-group
pillow
russian_names
easing-functions
numpy